renamed to europepmc_pipeline for consistency
